Move ISO function to helpers
